requirements-service.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-pptx==1.0.2
python-multipart==0.0.20
